LSU: Change encoding of 'sw' and limit LSU to one lane
if we execute two lw/sw insns in parallel we need extra Ports on dmem.
For now limit it to one lane to avoid congestions. Likewise for sw
encoding -- Before we used 'dst' as a read register which means that we
need an extra read port on the register file. This was not necessary as
we were using rs2 as an immediate. So simply switching rs2 and dst fixed
it.

Signed-off-by: Bastian Koppelmann <kbastian@mail.uni-paderborn.de>